Moves Handlers to custom namespace
